requirements.txt
langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain-community
faiss-cpu
langchain-core
langchain-classic
streamlit
langchain_groq